chore: consolidate postcss dependencies
