+++ packages/postcss/src/postcss-runtime.ts
@@ -1,0 +1,17 @@
+export type {
+  AcceptedPlugin,
+  AtRule,
+  Container,
+  Declaration,
+  Document,
+  Helpers,
+  Node,
+  Plugin,
+  PluginCreator,
+  ProcessOptions,
+  Processor,
+  Result,
+  Root,
+  Rule,
+} from 'postcss'
+export { default as postcss } from 'postcss'
