+++ packages/weapp-tailwindcss/package.json
@@ -213,9 +213,6 @@
     "lru-cache": "11.5.1",
     "magic-string": "0.30.21",
     "micromatch": "^4.0.8",
-    "postcss": "catalog:postcss85tilde",
-    "postcss-load-config": "^6.0.1",
-    "postcss-selector-parser": "^7.1.2",
     "semver": "~7.8.4",
     "tailwindcss-config": "workspace:*",
     "tailwindcss-patch": "catalog:tailwindcssPatch",
