+++ packages/weapp-tailwindcss/src/bundlers/shared/generator-css/legacy-compat.ts
@@ -2,7 +2,7 @@ import type { IStyleHandlerOptions } from '@weapp-tailwindcss/postcss/types'
 import type { TailwindResolvedSource } from '@/generator'
 import type { InternalUserDefinedOptions } from '@/types'
 import { readFileSync } from 'node:fs'
-import postcss from 'postcss'
+import { postcss } from '@weapp-tailwindcss/postcss'
 import { removeUnsupportedMiniProgramAtRules } from '../css-cleanup'
 import { removeTailwindSourceDirectives, resolveCssEntrySource } from './directives'
 import { collectDedupedPostTransformCompatCss, collectGeneratedSelectors, removeDuplicatedViteMarkers, removeGeneratedSelectorCompatCss } from './legacy-selectors'
