+++ packages/weapp-tailwindcss/src/bundlers/vite/official-tailwind-plugins.ts
@@ -1,42 +1,5 @@
 import type { Plugin } from 'vite'
 
-const tailwindPostcssPluginNames = new Set(['tailwindcss', '@tailwindcss/postcss'])
-
-export function getPostcssPluginName(plugin: unknown): string | undefined {
-  if (!plugin) {
-    return
-  }
-  if (typeof plugin === 'function' && 'postcss' in plugin) {
-    try {
-      return getPostcssPluginName(plugin())
-    }
-    catch {
-      return
-    }
-  }
-  if (typeof plugin !== 'object' || !('postcssPlugin' in plugin)) {
-    return
-  }
-  const { postcssPlugin } = plugin as { postcssPlugin?: unknown }
-  return typeof postcssPlugin === 'string' ? postcssPlugin : undefined
-}
-
-function isTailwindPostcssPlugin(plugin: unknown) {
-  const name = getPostcssPluginName(plugin)
-  return typeof name === 'string' && tailwindPostcssPluginNames.has(name)
-}
-
-export function removeTailwindPostcssPlugins(plugins: unknown[]) {
-  let removed = 0
-  for (let i = plugins.length - 1; i >= 0; i--) {
-    if (isTailwindPostcssPlugin(plugins[i])) {
-      plugins.splice(i, 1)
-      removed++
-    }
-  }
-  return removed
-}
-
 function isTailwindVitePlugin(plugin: unknown) {
   if (!plugin || typeof plugin !== 'object' || !('name' in plugin)) {
     return false
