+++ packages/weapp-tailwindcss/src/bundlers/webpack/loaders/weapp-tw-runtime-classset-loader.ts
@@ -2,7 +2,7 @@ import type webpack from 'webpack'
 import type { RuntimeLoaderWatchDependencies, WebpackRuntimeClassSetLoaderOptions } from './runtime-registry'
 import { Buffer } from 'node:buffer'
 import process from 'node:process'
-import postcss from 'postcss'
+import { postcss } from '@weapp-tailwindcss/postcss'
 import { removeUnsupportedCascadeLayers } from '@/tailwindcss/remove-unsupported-css'
 import { getWebpackLoaderRuntime } from './runtime-registry'
 import { registerWebpackWatchContext, registerWebpackWatchFile } from './watch-dependencies'
