+++ packages/weapp-tailwindcss/src/tailwindcss/v3-engine/source.ts
@@ -3,7 +3,7 @@ import type { TailwindV3ResolvedSource, TailwindV3SourceOptions } from './types'
 import type { TailwindcssPatcherLike } from '@/types'
 import path from 'node:path'
 import process from 'node:process'
-import postcss from 'postcss'
+import { postcss } from '@weapp-tailwindcss/postcss'
 import { loadConfig } from 'tailwindcss-config'
 import { resolveTailwindcssOptions } from '@/tailwindcss/patcher-options'
 import { omitUndefined } from '@/utils/object'
