+++ packages/weapp-tailwindcss/src/types/index.ts
@@ -1,7 +1,6 @@
 import type { ParseError, ParserOptions } from '@babel/parser'
-import type { CssPreflightOptions, IStyleHandlerOptions } from '@weapp-tailwindcss/postcss/types'
+import type { CssPreflightOptions, Document, IStyleHandlerOptions, Result as PostcssResult, Root } from '@weapp-tailwindcss/postcss'
 import type { SourceMap } from 'magic-string'
-import type { Document, Result as PostcssResult, Root } from 'postcss'
 import type { ILengthUnitsPatchOptions, TailwindcssPatcher } from 'tailwindcss-patch'
 import type { ICreateCacheReturnType } from '../cache'
 import type { ItemOrItemArray } from './base'
