+++ packages/weapp-tailwindcss/src/types/postcss-html-transform-shim.d.ts
@@ -1,5 +1,5 @@
 declare module '@weapp-tailwindcss/postcss/html-transform' {
-  import type { PluginCreator } from 'postcss'
+  import type { PluginCreator } from '@weapp-tailwindcss/postcss'
 
   export interface IOptions {
     /** 当前编译平台 */
